// src/app/services/user/currentUser.types.ts
export interface SipData {
  extension: string;
  sip_password: string;
  sip_server: string;
}

export interface CurrentUser
{
    id: string;
    username: string;
    email_address: string;
    first_name:  string;
    last_name:  string;
    title:  string;
    avatar?: string;
    active?: boolean;
    session_status: string;
    security_level: string;
    user_id: string;
    socket_id: string;
    sip_enabled: boolean;
    sipData?: SipData;
}